+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Rename Connect to Substra in code and ci files
 -   Change primary color from teal to blue
 -   Keep filtering/ordering setup when refreshing an asset list page
+-   Display all General customize columns by default
 
 ### Fixed